refactor: split gcode parser modules
