+++ test/gcode_import/gcode_import_parser_test.dart
@@ -142,6 +142,42 @@ void main() {
       expect(result.filamentLengthMm, 1500);
     });
 
+    test('parses named duration from streamed lines', () async {
+      final result = await parser.parseLineStream(
+        Stream.fromIterable([
+          '; generated by test tool 2.1',
+          '; TIME: 1h 2m 3s',
+          '; layer_height = 0.2',
+        ]),
+      );
+
+      expect(
+        result.estimatedDuration,
+        const Duration(hours: 1, minutes: 2, seconds: 3),
+      );
+      expect(result.layerHeightMm, 0.2);
+      expect(result.slicer, GCodeSlicer.unknown);
+      expect(
+        result.warnings.map((warning) => warning.code),
+        contains(GCodeParseWarningCode.unknownSlicer),
+      );
+    });
+
+    test('flags partial metadata when only some values exist', () {
+      final result = parser.parse('''
+; generated by test tool 2.1
+; TIME: 900
+; layer_height = 0.2
+''');
+
+      expect(result.estimatedDuration, const Duration(seconds: 900));
+      expect(result.layerHeightMm, 0.2);
+      expect(
+        result.warnings.map((warning) => warning.code),
+        contains(GCodeParseWarningCode.partialMetadata),
+      );
+    });
+
     test('keeps unknown slicer while extracting partial metadata', () {
       final result = parser.parse(load('unknown_basic.gcode'));
